+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Editor
   - Change call hierarchy to return selection range of usage, not function definition.
+  - Return `edits` in `codeAction/resolve` responses rather than `commands`. #655
 
 ## 2021.12.01-12.28.16